Move KeyboardFlags to dedicated keyboard module
